corr98

+++ cpp06/ex01/Serialization.hpp
@@ -30,8 +30,8 @@ class Serialization {
         Serialization& operator=(const Serialization& other); // Copy assignment operator
         ~Serialization();
     public:
-        static uintptr_t serialize(Data* ptr);
-        static Data* deserialize(uintptr_t raw);
+        static unsigned long serialize(Data* ptr);
+        static Data* deserialize(unsigned long raw);
 };
 
 #endif
